// src/api/generated/schema.ts
export interface components {
    schemas: {
// ...
        MusicReleaseDraft: {
            name: string;
            /** @enum {string} */
            releaseType: "album" | "ep" | "single" | "compilation" | "other";
            /** Format: date */
            releaseDate?: string | null;
            artistCredits: string[];
            image?: components["schemas"]["MusicReleaseDraftImage"] | null;
            tracks: components["schemas"]["MusicReleaseDraftTrack"][];
            spotifyReference?: components["schemas"]["SpotifyReference"] | null;
        };
// ...
    };
    responses: never;
    parameters: {
        ReleaseId: number;
        TrackId: number;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}